archon-main/pyproject.toml
[tool.poetry]
name = "sdss-archon"
version = "0.13.1a0"
description = "A library and actor to communicate with an STA Archon controller."
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/archon"
repository = "https://github.com/sdss/archon"
documentation = "https://sdss-archon.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "archon" }
]

[tool.poetry.dependencies]
python = "^3.9,<3.13"
sdsstools = "^1.5.2"
numpy = "^1.26.1"
sdss-clu = "^2.0.0b2"
click-default-group = "^1.2.2"
astropy = "^6.0"
fitsio = {version = "^1.2.1", optional = true}
nptyping = "^2.5.0"

[tool.poetry.group.dev.dependencies]
ipython = [
    {version = ">=8.0.0", python = ">=3.9,<3.10"},
    {version = "^8.17.2", python = ">=3.10"}
]
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = ">=3.3.1"
Sphinx = ">=3.0.0"
black = ">=20.8b1"
sphinx-click = ">=2.6.0"
sphinx-jsonschema = ">=1.16.7"
myst-parser = ">=0.14.0"
furo = ">=2021.6.18-beta.36"
nox = ">=2021.6.12"
sphinx-autobuild = ">=2021.3.14"
sphinx-copybutton = ">=0.3.3"
ruff = ">=0.0.284"

[tool.poetry.extras]
fitsio = ["fitsio"]

[tool.black]
line-length = 88
target-version = ['py311']
fast = true

[tool.ruff]
line-length = 88
target-version = 'py311'
select = ["E", "F", "I"]
exclude = ["__init__.pyi"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F403", "F401", "E402"]

[tool.ruff.isort]
known-first-party = ["ifoo"]
lines-after-imports = 2
section-order = ["future", "standard-library", "typing", "third-party", "sdss", "first-party", "local-folder"]

[tool.ruff.isort.sections]
typing = ["typing"]
sdss = ["sdsstools", "clu"]

[tool.pytest.ini_options]
addopts = "--cov archon --cov-report xml --cov-report html --cov-report term"
markers = [
    "commands: commands and replies for the test Archon"
]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
include = ["archon/*"]
omit = [
    "*/__init__.py"
]

[tool.coverage.report]
exclude_lines = [
    "# pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING"
]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
